Reuse the common functions
