+++ zephyr/common/cloud/endpoints/__init__.py
@@ -1,0 +1,13 @@
+from .endpoints import (TestCaseEndpoints,
+                        TestCycleEndpoints,
+                        TestPlanEndpoints,
+                        TestExecutionEndpoints,
+                        FolderEndpoints,
+                        StatusEndpoints,
+                        PriorityEndpoints,
+                        EnvironmentEndpoints,
+                        ProjectEndpoints,
+                        LinkEndpoints,
+                        IssueLinksEndpoints,
+                        AutomationEndpoints,
+                        HealthcheckEndpoints)
