+++ zephyr/scale/cloud/endpoints/__init__.py
@@ -1,13 +1,18 @@
-from .endpoints import (TestCaseEndpoints,
-                        TestCycleEndpoints,
-                        TestPlanEndpoints,
-                        TestExecutionEndpoints,
-                        FolderEndpoints,
-                        StatusEndpoints,
-                        PriorityEndpoints,
-                        EnvironmentEndpoints,
-                        ProjectEndpoints,
-                        LinkEndpoints,
-                        IssueLinksEndpoints,
-                        AutomationEndpoints,
-                        HealthcheckEndpoints)
+from .endpoints import (
+    TestCaseEndpoints,
+)
+
+from zephyr.common.cloud.endpoints.endpoints import(
+    TestCycleEndpoints,
+    TestPlanEndpoints,
+    TestExecutionEndpoints,
+    FolderEndpoints,
+    StatusEndpoints,
+    PriorityEndpoints,
+    EnvironmentEndpoints,
+    ProjectEndpoints,
+    LinkEndpoints,
+    IssueLinksEndpoints,
+    AutomationEndpoints,
+    HealthcheckEndpoints
+)
