+++ zephyr/scale/server/endpoints/endpoints.py
@@ -1,4 +1,4 @@
-from zephyr.scale.zephyr_scale_session import ZephyrScaleSession
+from zephyr.common.cloud.zephyr_cloud_session import ZephyrScaleSession
 from .paths import ServerPaths as Paths
