+++ zephyr/squad/__init__.py
@@ -1,1 +1,1 @@
-from zephyr.squad.squad import ZephyrSquad
+from zephyr.squad.squad import API_V1, API_V2, ZephyrSquad
